Add lxml to requirements.txt
In splitting the tempest plugin code out from the keystone repository we
copied over everything from the keystone_tempest_plugin/ directory but
forgot to copy the shared requirements living in the main repo. This
patch adds the lxml library to the plugin's requirements so that the
plugin can check SAML things.

Change-Id: I47f59505126509b6b52cf04b8f7501e8ef4041b7