Make checking for singular SPs in list instead of list equality
There is a race condition when the test_service_providers_in_token
test is run at the same time as the k2k test because an extra SP
will appear in the list.

By checking items in the list individually instead of comparing
list equality this should fix the issue.

Change-Id: I13a7a747e108562b326aee1b88485a377530f8a5