Update supported branches
This patch updates the jobs to reflect the latest suported branches for
keystone.

Test jobs for both antleope (2023.1) and bobcat (2023.2) have been
added, and the jobs for the xena and yoga branches have been removed as
they are no longer maintained.

This patch also makes the protection jobs non-voting as they are
expected to fail due to policy changes in keystone.  A follow-up patch
fixes the test and re-enables the job.

Change-Id: I2d3968672eb4dd32a163827a7e24384578a4c913